ash: avoid a warning about a cast
It is incorrect, of course, to cast a `const char *` to a `char *`, but
some functions are simply declared with incorrect types (such as
execve(), which is a system function, therefore tryexec() cannot do much
but follow suite).

This is a companion patch to 57e14f870 (ash: skip PATH if executable and
command are both busybox, 2015-06-02).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>