mingw: use UTF-16 variant to spawn processes
As every Win32 program, BusyBox-w32's use of spawnve() implies that the
command-line arguments and the environment are dependent on the active
code page.

In Git for Windows' context, we want to treat all command-line arguments
and the environment variables/values as UTF-8, though. Let's simply
convert them to the wide character versions and use _wspawnve() instead.

While at it, let's use mingw_pathconv() to benefit from the path
translation, too.

Note: while developing this patch, an alternative approach was tested
based on CreateProcessW() instead of _wspawnve(). However, the semantics
of these two API calls are different in at least one subtle way: while
file handles are inherited when passing bInheritHandles = TRUE to
CreateProcessW(), file *descriptors* are only inherited if they happen
to refer to stdin/stdout/stderr. In contrast, _wspawnve() inherits also
file descriptors > 2, which is something that our --forkshell support
relies on when opening pipes. So _wspawnve() it is.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>