ash: when updating the pwd, give mingw_pathconv() a chance
The newly-introduced mingw_pathconv() function attempts to replicate
MSYS2's auto-conversion from /usr/bin to <MSYS2_ROOT>\usr\bin (where
MSYS2_ROOT is relative to the executable, determined on-the-fly) and
from /c/WINDOWS to c:\WINDOWS.

It won't work, though, if we prefix those paths with the drive already,
though. So let's not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>