kill: use gentler method than TerminateProcess()
As Git for Windows' source code recently learned, let's also avoid using
TerminateProcess() in BusyBox-w32: it does not allow the killed
processes' atexit() handlers to run.

Instead, jump through a couple of hoops by injecting a remote thread
that executes the ExitProcess() function.

This allows the atexit() handlers to run, at which point the exit code
of the process can already be queried via GetExitCodeProcess(), and
appropriate action can be taken, such as killing child processes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>