mingw: try to kill spawned processes when killed by a signal
Our new behavior for the `kill` command is to inject a thread that runs
the ExitProcess() function with an exit code of the form `128 + signal`.

This means that we can sort of emulate POSIX signaling behavior by
terminating the child processes with the very same signal.

Let's do that.

As a special quirk, we need to roll the _P_WAIT functionality of
spawnve() ourselves now, because we would otherwise not have a chance to
get the handle of the spawned process to kill.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>