Fix potential use-after-free
The getenv() function is inherently thread-unsafe: the returned pointers
are potentially invalidated whenever setenv() or unsetenv() is called.

This is not just some theoretical consideration: immediately after
obtaining the pointer to the BUSYBOX_APPLET_NAME variable's value, we
call unsetenv() for the very same environment variable.

Oddly enough, this invalidates the pointer only occasionally, and then
usually by resetting the value to the empty string. The symptom is a
seemingly random:

	: applet not found

Let's just strdup() the string; We do not even need to bother with
free()ing it because we allocate this in the main() function and the
process will simply exit after using the value in run_applet_and_exit(),
anyway.

This fixes a bug introduced in 7af304a47 (win32: appletlib: special way
to recognize applet name on Windows, 2010-04-05).

That commit also seems to purport that argv[0] *must* be the executable
path on Windows when spawning new processes. That is incorrect, however,
possibly making the entire BUSYBOX_APPLET_NAME workaround moot. But
that's a story for another day.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>