dump: help gcc understand that there is no problem
GCC 7.1.0 complains about this:

libbb/dump.c: In function 'bb_dump_dump':
libbb/dump.c:486:17: warning: '%03o' directive writing between 3 and 11 bytes into a region of size 10 [-Wformat-overflow=]
   sprintf(buf, "%03o", (int) *p);
                 ^~~~
libbb/dump.c:486:16: note: directive argument in the range [0, 2147483647]
   sprintf(buf, "%03o", (int) *p);
                ^~~~~~
libbb/dump.c:486:3: note: 'sprintf' output between 4 and 12 bytes into a destination of size 10
   sprintf(buf, "%03o", (int) *p);
   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The reason is that an int really has that range, and `buf` really is too
small for arbitrary ints.

However, *p refers to an unsigned char, so we already know that the
output will occupy between 2 and 4 bytes including the trailing NUL.

GCC seems to be happy to recognize the correct range, though, if we cast
to an *unsigned* int. So let's do just that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>