Silence misguided GCC warning about alignment issues
When compiling xz_dec_stream.c with GCC 7.1.0, it complains thusly:

	In function 'dec_stream_footer':
	error: dereferencing type-punned pointer will break strict-aliasing
	      rules [-Werror=strict-aliasing]
	   if (xz_crc32(s->temp.buf + 4, 6, 0) != get_le32(s->temp.buf))
	       ^~

The thing is, the `buf` field was put just after two fields of type
size_t for the express purpose of avoiding alignment issues, as per the
comment above the `temp` struct.

Meaning: GCC gets this all wrong and should not complain. So let's force
it to be quiet for a couple of lines.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>