mingw: clarify that waitpid(-1, ...) is not supported
When passing -1 as pid to the waitpid() function, it is supposed to wait
for *any* child to exit. That is a bit tough to emulate on Windows
given that waitpid() returns only one pid, and if multiple children have
exited, subsequent calls to waitpid() should produce all of their pids
subsequently. Oh, and we would have to figure out which processes were
spawned (or remember that list, just like we do it for
kill_children_on_signal() but with semantics *just* different enough so
that we cannot reuse that logic).

So let's punt and simply state that we do not support that
functionality.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>