mingw: special-case xargs -P on Windows
It is pretty hard to emulate the waitpid() semantics on Windows when pid
is -1, i.e. when waiting for any child to exit.

The problem is not so much accumulating the process handles of children
we spawned (we do that already to emulate killing processes including
their children), but the fact that waitpid() returns exactly one pid
even when multiple children have exited, retaining the rest of them for
subsequent waitpid() calls. The existing code to retain handles of
spawned children, however, does *not* want to retain those handles but
do away with them as soon as it is detected that the processes exited.

Naturally, we could introduce a second array of process handles just for
waitpid() but the question is how we would release those handles if
waitpid() is never called.

A much more elegant way is to add Windows-specific code to xargs.c that
specifically handles the child processes spawned by xargs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>