libbb: exclude more Linux-specific functions from being compiled
In 7a4491e1f (libbb: don't compile various Linux-specific functions,
2017-08-24), we started excluding a number of Linux-specific functions
that handle things like SELinux, ioctls, devices and forking.

But we forgot a couple functions that depend on the ones we now
excluded. This leads to compile failures with -Werror unless
cross-compiling on Linux, due to undefined symbols.

To fix this, exclude those functions which depend on the ones that were
excluded (they do not have any callers on Windows anyway, so all is
good).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>