mingw: avoid heap corruption
After calling putenv(), there is no guarantee that a previously-valid
`environ` is still valid. Since we loop over putenv()s, and since the
loop condition is `environ && *environ`, we are really not safe.

So unsafe, in fact, that at least one heap corruption could be
attributed to it in this developer's setup.

Let's fix it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>