ash: duplicate proc handle after mingw_spawn_proc
Do `DuplicateHandle()` after calling `mingw_spawn_proc()` before
calling `WaitForMultipleObjects()` in case the process handle is
probable to be closed in `cull_exited_processes()` in process.c.

Note that in `spawn_forkshell()`, `ret` is an `intptr_t` returned
from `mingw_spawn_proc()`, and can be cast to a `HANDLE`, which
is exactly the process handle. Later in the function it is passed
to `forkparent()`, where it is stored in a `struct procstat` as
`ps->ps_proc`. After that, the handle is used in
`WaitForMultipleObjects()` in `waitpid_child()`.

Consider the situation below:

1. `evalpipe()` is called so that two `spawn_forkshell()` is to
    be done
2. The first child process is created by `mingw_spawn_proc()` in
   `spawn_forkshell()`, and its process handle is returned
3. The first child process gets its timeslice and finishes running
   quickly (though normally this is not likely to happen)
4. The second child process is created by `mingw_spawn_proc()`, in
   which there are several chances to call `cull_exited_processes()`,
   where the first child's handle is closed
5. Backing to ash, in `waitpid_child()`, `WaitForMultipleObjects()`
   is called with handles of the two child processes. But the first
   one's handle is now invalid
6. The return value(`idx`) is thus invalid, then `waitpid_child()`
   returns -1 and `dowait()` returns -1
7. Because of the lack of error handling of `dowait()`, it is called
   repeatedly inside a while loop in `waitforjob()`.
8. Busybox gets into an endless loop.

This issue is found during the development of a tool that injects
a DLL into one process's all descandant processes. This DLL works
by hooking `CreateProcessW` call, forcing the hooked process to also
inject the child process it just created. This will change the thread
scheduling behaviour a little bit among the hooked process and its
child processes, which uncovers this problem.

To resolve, let's `DuplicateHandle()` in `spawn_forkshell()`, after
`mingw_spawn_proc()`; and properly close the duplicated handle.